support Continue Generate
