+++ src/locales/default/chat.ts
@@ -188,6 +188,7 @@ export default {
   },
 
   messageAction: {
+    continueGeneration: '继续生成',
     delAndRegenerate: '删除并重新生成',
     deleteDisabledByThreads: '存在子话题，不能删除',
     regenerate: '重新生成',
